requirements.txt
requests==2.31.0
python-nmap==0.7.1
beautifulsoup4==4.12.0
